# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger Api with Laravel Sanctum authentication",
        "version": "0.1"
    },
    "paths": {
        "/api/login/store": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Authenticate User Login and generate token",
                "operationId": "65b4746d608da585bf08d8508bed8c54",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a list of projects",
                "operationId": "2ea4919d0a733012d5e30a9c60233aca",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Project name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Project status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Project deadline date",
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get lists of projects"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/projects/store": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Store projects",
                "operationId": "8517397fe2b0f9879a11d6298e5a7174",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Project Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Project Description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Project status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "active or inactive"
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Project deadline date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "YYYY-MM-DD"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Created Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized request"
                    },
                    "403": {
                        "description": "Forbidden request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/projects/{projectId}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "View single project",
                "operationId": "viewProjectId",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project View Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show single project"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized request"
                    },
                    "403": {
                        "description": "Forbidden request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update single project",
                "operationId": "updateProjectId",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project Update Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Project Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Project Description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Project status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Project deadline date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Updated Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized request"
                    },
                    "403": {
                        "description": "Forbidden request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Remove single projects",
                "operationId": "deleteProjectId",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "query",
                        "description": "Project Remove Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Remove Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Unauthorized request"
                    },
                    "403": {
                        "description": "Forbidden request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Laravel sanctum oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000//oauth/authorize",
                        "tokenUrl": "http://localhost:8000//oauth/token",
                        "refreshUrl": "http://localhost:8000//token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}